lol (#17)
* lol

* simplify serch

* auto save / load is back

* cleanup

* simplify

* fix save / load states (ish)

+++ Delta/Views/Context Menus/Game.swift
@@ -2,17 +2,8 @@ import SwiftUI
 
 struct GameContextMenu: View {
     @Environment(\.managedObjectContext) var context
-    @EnvironmentObject var currentlyPlaying: CurrentlyPlayingStore
     
     @ObservedObject var game: GameEntity
-    
-    @State private var isShowingAddToPlaylist = false
-    @State private var isShowingConfirmRemove = false
-    @State private var isShowingSaveStates = false
-
-    func play() {
-        currentlyPlaying.selectedGame(game)
-    }
     
     func toggleFave() {
         game.objectWillChange.send()
@@ -20,14 +11,10 @@ struct GameContextMenu: View {
         try? game.managedObjectContext?.save()
     }
     
-    func toggleConfirmRemove() {
-        isShowingConfirmRemove.toggle()
-    }
-    
     var body: some View {
         Group {
             if game.hasROM {
-                Button(action: play) {
+                Button(action: ActionCreator().presentEmulator(self.game)) {
                     HStack {
                         Text("Play")
                         Spacer()
@@ -61,49 +48,36 @@ struct GameContextMenu: View {
                     }
                 }
             }
-            Button(action: {
-                self.isShowingAddToPlaylist.toggle()
-            }) {
+            Button(action: ActionCreator().presentAddToPlaylist(self.game)) {
                 HStack {
                     Text("Add to Playlist")
                     Spacer()
                     Image(systemSymbol: .textBadgePlus)
                 }
             }
-            .sheet(isPresented: $isShowingAddToPlaylist) {
-                AddToPlaylist(isShowing: self.$isShowingAddToPlaylist, game: self.game)
-                    .environment(\.managedObjectContext, self.context)
-            }
             if game.hasROM {
-                Button(action: {
-                    self.isShowingSaveStates.toggle()
-                }) {
+                Button(action: ActionCreator().presentSavedStates(self.game)) {
                     HStack {
                         Text("View Save States")
                         Spacer()
                         Image(systemSymbol: .moon)
                     }
                 }
-                .sheet(isPresented: $isShowingSaveStates) {
-                    SaveStatesView(game: self.game) {
-                        self.isShowingSaveStates = false
-                        self.currentlyPlaying.selectedSave($0)
-                    }
+            }
+            Button(action: ActionCreator().presentLookup(self.game)) {
+                HStack {
+                    Text("Search Web")
+                    Spacer()
+                    Image(systemSymbol: .globe)
                 }
-                Button(action: toggleConfirmRemove) {
+            }
+            if game.hasROM {
+                Button(action: ActionCreator().presentRemoveFromLibraryConfirmation(self.game)) {
                     HStack {
                         Text("Delete from Library")
                         Spacer()
                         Image(systemSymbol: .trash)
                     }
-                }
-                .actionSheet(isPresented: $isShowingConfirmRemove) {
-                    ActionSheet(title: Text("Are you sure you want to remove this game from your library?"), message: nil, buttons: [
-                        .destructive(Text("Delete Game")) {
-                            self.game.deleteFromLibrary()
-                        },
-                        .cancel()
-                        ])
                 }
             }
         }
